--amend

+++ wf_reg_test/__main__.py
@@ -134,7 +134,8 @@ def regenerate() -> None:
 def clear() -> None:
     (data_path / "nf-core_executions.yaml").write_text("[]")
     (data_path / "snakemake-workflow-catalog_executions.yaml").write_text("[]")
-    shutil.rmtree(".repos")
+    if Path(".repos").exists():
+        shutil.rmtree(".repos")
 
 
 @main.command()
